Decoupling JPA Crud repository
In this PR, I'm adding a JpaExpensesRepository that is acting as an
internal adapter for the CrudRepository interface from JPA. Also, I'm
adding an ExpensesRepo interface in the entities folder (this folder will be renamed later) which will be used by the controller and maybe by the domain later on. This approach allows me to decouple my domain from the external implementation of the database/framework.